requirements.txt
streamlit
openai
pandas
sendgrid
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib